feat: seed movie, talent and genre permissions and protect routes.

+++ src/routes/v1/dashboard/genre.route.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import GenreController from "../../../controllers/dashboard/genre.controller";
 import { controllerAsync } from "../../../middlewares/handlers/controllerAsync.handler";
+import { checkPermission } from "../../../middlewares/permission.middleware";
 import GenreRepository from "../../../repositories/genre.repository";
 import GenreService from "../../../services/dashboard/genre.service";
 
@@ -12,11 +13,13 @@ const gerneController = new GenreController(genreService);
 router
 	.route("/")
 	.post([
+		checkPermission("genre", "create"),
 		controllerAsync((request, response) =>
 			gerneController.createGenre(request, response),
 		),
 	])
 	.get([
+		checkPermission("genre", "read"),
 		controllerAsync((request, response) =>
 			gerneController.getAllGenres(request, response),
 		),
@@ -25,11 +28,13 @@ router
 router
 	.route("/:id")
 	.put([
+		checkPermission("genre", "update"),
 		controllerAsync((request, response) =>
 			gerneController.updateGenre(request, response),
 		),
 	])
 	.delete([
+		checkPermission("genre", "delete"),
 		controllerAsync((request, response) =>
 			gerneController.deleteGenre(request, response),
 		),
