+++ src/routes/v1/dashboard/movie.route.ts
@@ -4,6 +4,7 @@ import {
 	controllerAsync,
 	controllerAsyncWithTransaction,
 } from "../../../middlewares/handlers/controllerAsync.handler";
+import { checkPermission } from "../../../middlewares/permission.middleware";
 import { uploadImage } from "../../../middlewares/upload.middleware";
 import MovieRepository from "../../../repositories/movie.repository";
 import MovieService from "../../../services/dashboard/movie.service";
@@ -16,11 +17,13 @@ const movieController = new MovieController(movieService);
 router
 	.route("/")
 	.get([
+		checkPermission("movie", "read"),
 		controllerAsync((request, response) =>
 			movieController.getAllMovies(request, response),
 		),
 	])
 	.post([
+		checkPermission("movie", "create"),
 		uploadImage.single("image"),
 		controllerAsyncWithTransaction((request, response, session) =>
 			movieController.createMovie(request, response, session),
@@ -30,17 +33,20 @@ router
 router
 	.route("/:id")
 	.get([
+		checkPermission("movie", "read"),
 		controllerAsync((request, response) =>
 			movieController.getSingleMovie(request, response),
 		),
 	])
 	.put([
+		checkPermission("movie", "update"),
 		uploadImage.single("image"),
 		controllerAsyncWithTransaction((request, response, session) =>
 			movieController.updateMovie(request, response, session),
 		),
 	])
 	.delete([
+		checkPermission("movie", "delete"),
 		controllerAsyncWithTransaction((request, response, session) =>
 			movieController.deleteMovie(request, response, session),
 		),
